// src/utils/types/SensoryInputs.ts
export interface SensoryInputs {
  age: number
  random: number
  currentPositionY: number
  currentPositionX: number
  generation: number
  sizeOfMapX: number
  sizeOfMapY: number
  population: number
  distanceOfNextObjectNorth: number
  distanceOfNextObjectEast: number
  distanceOfNextObjectSouth: number
  distanceOfNextObjectWest: number
}